services/auth_service/requirements.txt
Flask==3.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
pytest==7.4.3
pytest-cov==4.1.0
flasgger==0.9.7.1
requests==2.31.0

